update highcharts in views
I am submitting one more pull request consideration, hope it helps. The
highcharts javascript code we pulled into a partial tonight had query
calls within the source.  I've pulled those out and also moved the code
for setup of the data in the controller into the models.  This was the
same pattern I mentioned earlier tonight that may make it easier to test
later.  This should hopefully cut down on the querying and time required
for each individual call.